NDS-1138: create basic html table component
