+++ components/src/RVTable/index.js
@@ -1,0 +1,1 @@
+export { default as Table } from "./Table";
